tilt: add --verbose, --debug flags to synclet
